hide redundant information for token transfer transactions
